requirement.txt
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1
PyYAML==6.0.1
numpy==1.24.4
timm==0.3.2
tensorboard==2.14.0
einops==0.7.0
pandas==2.0.3
matplotlib==3.7.4
tqdm==4.66.1
scipy==1.10.1
scikit-learn==1.3.2
ninja==1.11.1.1
packaging==23.2
flash-attn==2.5.5